+++ src/main/java/org/prgms/locomocoserver/categories/domain/CategoryType.java
@@ -1,0 +1,5 @@
+package org.prgms.locomocoserver.categories.domain;
+
+public enum CategoryType {
+    USER, MOGAKKO
+}
