--- src/main/java/org/prgms/locomocoserver/global/categories/domain/CategoryType.java
@@ -1,5 +1,0 @@
-package org.prgms.locomocoserver.global.categories.domain;
-
-public enum CategoryType {
-    USER, MOGAKKO
-}
