+++ src/main/java/org/prgms/locomocoserver/tags/domain/Tag.java
@@ -1,4 +1,4 @@
-package org.prgms.locomocoserver.global.tags.domain;
+package org.prgms.locomocoserver.tags.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -13,7 +13,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.prgms.locomocoserver.global.categories.domain.Category;
+import org.prgms.locomocoserver.categories.domain.Category;
 
 @Entity
 @Getter
